refactor: apply comments for layout
